api: split utils.cpp into multiple implementation files
