+++ test/geometry/meson.build
@@ -1,6 +1,6 @@
 geometry_test = executable(
     'geometry_test',
     'geometry_test.cpp',
-    dependencies: [wfconfig, doctest, libwayfire],
+    dependencies: mocklib,
     install: false)
 test('Geometry test', geometry_test)
